+++ pyobjc-core/Modules/objc/ctests.h
@@ -1,6 +1,10 @@
 #ifndef PyObjC_CTESTS_H
 #define PyObjC_CTESTS_H
 
+NS_ASSUME_NONNULL_BEGIN
+
 extern int PyObjC_init_ctests(PyObject*);
+
+NS_ASSUME_NONNULL_END
 
 #endif /* PyObjC_CTESTS_H */