+++ pyobjc-core/Modules/objc/helpers-foundation-nsobject.m
@@ -3,111 +3,12 @@
  */
 #include "pyobjc.h"
 
-/* Temporary code: used to get a baseline for attempts to reduce the overhead
- * of calling ObjC methods. This direct call of the description should be as
- * close to optimal as we'll be able to get the generic code.
- *
- * First tests appear to indicate that there are significant improvements on
- * the table (20% to 30% change in method call overhead)
- */
-#undef PyObjC_EXPLICT_DESCRIPTION_HELPER
+NS_ASSUME_NONNULL_BEGIN
 
-#ifdef PyObjC_EXPLICT_DESCRIPTION_HELPER
-static PyObject*
-call_NSObject_description(PyObject* method, PyObject* self,
-                          PyObject* const* arguments __attribute__((__unused__)),
-                          size_t           nargs)
-{
-    id                result = nil;
-    struct objc_super spr;
-    IMP               anIMP;
-    NSObject*         anObject;
-    SEL               aSel;
-
-    if (PyObjC_CheckArgCount(method, 0, 0, nargs) == -1)
-        return NULL;
-
-    if (unlikely(PyObjCIMP_Check(method))) {
-        anIMP    = PyObjCIMP_GetIMP(method);
-        anObject = PyObjCObject_GetObject(self);
-        aSel     = PyObjCIMP_GetSelector(method);
-
-        Py_BEGIN_ALLOW_THREADS
-            @try {
-                result = ((id(*)(id, SEL))anIMP)(anObject, aSel);
-
-            } @catch (NSObject* localException) {
-                PyObjCErr_FromObjC(localException);
-                result = nil;
-            }
-        Py_END_ALLOW_THREADS
-
-    } else {
-        objc_superSetReceiver(spr, PyObjCObject_GetObject(self));
-        objc_superSetClass(spr, PyObjCSelector_GetClass(method));
-        aSel = PyObjCSelector_GetSelector(method);
-
-        Py_BEGIN_ALLOW_THREADS
-            @try {
-                result = ((id(*)(struct objc_super*, SEL))objc_msgSendSuper)(&spr, aSel);
-
-            } @catch (NSObject* localException) {
-                PyObjCErr_FromObjC(localException);
-                result = nil;
-            }
-        Py_END_ALLOW_THREADS;
-    }
-
-    if (unlikely(result == nil && PyErr_Occurred())) {
-        return NULL;
-    }
-
-    return PyObjCObject_New(result, 0, YES);
-}
-
-static void
-imp_NSObject_description(ffi_cif* cif __attribute__((__unused__)), void* resp,
-                         void** args __attribute__((__unused__)), void* callable)
-{
-    int       err;
-    PyObject* v      = NULL;
-    PyObject* result = NULL;
-
-    PyObjC_BEGIN_WITH_GIL
-        v = PyObjC_IdToPython(*(id*)args[0]);
-        if (unlikely(v == NULL)) {
-            Py_DECREF(arglist);
-            PyObjC_GIL_FORWARD_EXC();
-        }
-        v = PyObjC_AdjustSelf(v);
-        if (unlikely(v == NULL)) {
-            Py_DECREF(arglist);
-            PyObjC_GIL_FORWARD_EXC();
-        }
-
-        PyObject* args[2] = {NULL, v};
-        result            = PyObject_Vectorcall(callable, args + 1,
-                                                1 | PY_VECTORCALL_ARGUMENTS_OFFSET, NULL);
-        if (unlikely(result == NULL)) {
-            Py_DECREF(v);
-            PyObjC_GIL_FORWARD_EXC();
-        }
-        Py_DECREF(v);
-
-        err = depythonify_c_value(@encode(id), result, resp);
-        Py_DECREF(result);
-        if (unlikely(err == -1)) {
-            PyObjC_GIL_FORWARD_EXC();
-        }
-
-    PyObjC_END_WITH_GIL
-}
-
-#endif /* PyObjC_EXPLICT_DESCRIPTION_HELPER */
-
-static PyObject*
-call_NSObject_alloc(PyObject* method, PyObject* self,
-                    PyObject* const* arguments __attribute__((__unused__)), size_t nargs)
+static PyObject* _Nullable call_NSObject_alloc(PyObject* method, PyObject* self,
+                                               PyObject* const* arguments
+                                               __attribute__((__unused__)),
+                                               size_t nargs)
 {
     id                result = nil;
     struct objc_super spr;
@@ -200,10 +101,10 @@ imp_NSObject_alloc(ffi_cif* cif __attribute__((__unused__)), void* resp,
     PyObjC_END_WITH_GIL
 }
 
-static PyObject*
-call_NSObject_dealloc(PyObject* method, PyObject* self,
-                      PyObject* const* arguments __attribute__((__unused__)),
-                      size_t           nargs)
+static PyObject* _Nullable call_NSObject_dealloc(PyObject* method, PyObject* self,
+                                                 PyObject* const* arguments
+                                                 __attribute__((__unused__)),
+                                                 size_t nargs)
 {
     struct objc_super spr;
     IMP               anIMP;
@@ -297,10 +198,10 @@ imp_NSObject_dealloc(ffi_cif* cif __attribute__((__unused__)),
     PyObjC_END_WITH_GIL
 }
 
-static PyObject*
-call_NSObject_release(PyObject* method, PyObject* self,
-                      PyObject* const* arguments __attribute__((__unused__)),
-                      size_t           nargs)
+static PyObject* _Nullable call_NSObject_release(PyObject* method, PyObject* self,
+                                                 PyObject* const* arguments
+                                                 __attribute__((__unused__)),
+                                                 size_t nargs)
 {
     struct objc_super spr;
     IMP               anIMP;
@@ -354,9 +255,10 @@ call_NSObject_release(PyObject* method, PyObject* self,
     return Py_None;
 }
 
-static PyObject*
-call_NSObject_retain(PyObject* method, PyObject* self,
-                     PyObject* const* arguments __attribute__((__unused__)), size_t nargs)
+static PyObject* _Nullable call_NSObject_retain(PyObject* method, PyObject* self,
+                                                PyObject* const* arguments
+                                                __attribute__((__unused__)),
+                                                size_t nargs)
 {
     struct objc_super spr;
     IMP               anIMP;
@@ -487,13 +389,6 @@ PyObjC_setup_nsobject(void)
 {
     int r;
 
-#ifdef PyObjC_EXPLICT_DESCRIPTION_HELPER
-    r = PyObjC_RegisterMethodMapping(objc_lookUpClass("NSObject"), @selector(description),
-                                     call_NSObject_description, imp_NSObject_description);
-    if (r != 0)
-        return r;
-#endif /* PyObjC_EXPLICT_DESCRIPTION_HELPER */
-
     r = PyObjC_RegisterMethodMapping(objc_lookUpClass("NSObject"), @selector(alloc),
                                      call_NSObject_alloc, imp_NSObject_alloc);
     if (r != 0)
@@ -516,3 +411,5 @@ PyObjC_setup_nsobject(void)
 
     return r;
 }
+
+NS_ASSUME_NONNULL_END